Fix bug when equals used and style changes
Fixed a bug whereby after the use of "=" if the user then changes the number display the equals operator could be used again as the program. This should not be the case as if any changes are made this is essentially a new number and operation and we need to assign the new number to number A and go from there.